implemented :list and :list :eval for the widget tree language
expand-child has been renamed to expand-next because it could return
multiple new forms, expand-compound-widget has been adapted to really
do this if necessary. layout-expand-children thus does no longer
encapsulate the new child(ren) with a (list ...).

There is a new 5am-Testsuite in there. It is not loaded if *features*
does not contain '5am.